packages/agent-resources/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-resources"
version = "0.2.0"
description = "Add Claude Code skills, commands, and agents from GitHub to your project"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "httpx>=0.27",
    "typer>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
skill-add = "agent_resources.cli.skill:app"
command-add = "agent_resources.cli.command:app"
agent-add = "agent_resources.cli.agent:app"
create-agent-resources-repo = "agent_resources.cli.create:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_resources"]
